+++ tremor-pipeline/src/op/trickle/select.rs
@@ -42,7 +42,22 @@ pub struct GroupData<'groups> {
     window: WindowImpl,
     aggrs: Aggregates<'groups>,
     id: EventId,
+    transactional: bool,
 }
+
+impl<'groups> GroupData<'groups> {
+    fn track_transactional(&mut self, transactional: bool) {
+        self.transactional = self.transactional || transactional;
+    }
+
+    fn reset(&mut self) {
+        for aggr in &mut self.aggrs {
+            aggr.invocable.init();
+        }
+        self.transactional = false;
+    }
+}
+
 type Groups<'groups> = HashMap<String, GroupData<'groups>>;
 rental! {
     pub mod rentals {
@@ -509,9 +524,10 @@ fn execute_select_and_having(
     local_stack: &LocalStack,
     state: &Value<'static>,
     env: &Env,
-    event_id_gen: &mut EventIdGenerator,
+    event_id: EventId,
     event: &Event,
     origin_uri: Option<EventOriginUri>,
+    transactional: bool,
 ) -> Result<Option<(Cow<'static, str>, Event)>> {
     let (event_payload, event_meta) = event.data.parts();
 
@@ -535,17 +551,17 @@ fn execute_select_and_having(
             .into());
         }
     }
-    let mut event_id = event_id_gen.next_id();
-    event_id.track(&event.id);
     Ok(Some((
         OUT,
         Event {
             id: event_id,
             ingest_ns: event.ingest_ns,
-            // TODO avoid origin_uri clone here
             origin_uri,
+            // TODO: this will ignore op_metas from all other events this one is based upon and might break operators requiring this
+            op_meta: event.op_meta.clone(),
             is_batch: false,
             data: (result.into_static(), event_meta.clone_static()).into(),
+            transactional,
             ..Event::default()
         },
     )))
@@ -563,6 +579,9 @@ fn accumulate(
 ) -> Result<()> {
     // we incorporate the event into this group below, so track it here
     group.id.track(&event.id);
+
+    // track transactional state for the given event
+    group.transactional = group.transactional || event.transactional;
 
     let (event_data, event_meta) = event.data.parts();
     for aggr in &mut group.aggrs {
@@ -610,6 +629,7 @@ fn get_or_create_group<'window>(
                     aggrs,
                     group: group_value.clone_static(),
                     id: idgen.next_id(), // after all this is a new event
+                    transactional: false,
                 }
             });
             e.insert(k, v)
@@ -795,9 +815,10 @@ impl Operator for TrickleSelect {
                         &local_stack,
                         state,
                         &env,
-                        &mut self.event_id_gen,
+                        event.id.clone(),
                         &event,
                         event.origin_uri.clone(),
+                        event.transactional // no windows, we can safely pass through the events field
                     )) {
                         events.push(port_and_event);
                     };
@@ -825,7 +846,6 @@ impl Operator for TrickleSelect {
                         &group_str,
                         &group_value,
                     ));
-
                     let window_event = stry!(this_group.window.on_event(&event));
                     if window_event.emit && !window_event.include {
                         // push
@@ -836,6 +856,8 @@ impl Operator for TrickleSelect {
                             meta: &node_meta,
                             recursion_limit: tremor_script::recursion_limit(),
                         };
+                        let mut outgoing_event_id = self.event_id_gen.next_id();
+                        mem::swap(&mut outgoing_event_id, &mut this_group.id);
                         if let Some(port_and_event) = stry!(execute_select_and_having(
                             &stmt,
                             &node_meta,
@@ -843,16 +865,16 @@ impl Operator for TrickleSelect {
                             &local_stack,
                             state,
                             &env,
-                            &mut self.event_id_gen,
+                            outgoing_event_id,
                             &event,
                             None,
+                            this_group.transactional
                         )) {
                             events.push(port_and_event);
                         };
                         // re-initialize aggr state for new window
-                        for aggr in &mut this_group.aggrs {
-                            aggr.invocable.init();
-                        }
+                        // reset transactional state for outgoing events
+                        this_group.reset();
                     }
 
                     // accumulate
@@ -884,6 +906,9 @@ impl Operator for TrickleSelect {
                             meta: &node_meta,
                             recursion_limit: tremor_script::recursion_limit(),
                         };
+
+                        let mut outgoing_event_id = self.event_id_gen.next_id();
+                        mem::swap(&mut outgoing_event_id, &mut this_group.id);
                         if let Some(port_and_event) = stry!(execute_select_and_having(
                             &stmt,
                             &node_meta,
@@ -891,16 +916,16 @@ impl Operator for TrickleSelect {
                             &local_stack,
                             state,
                             &env,
-                            &mut self.event_id_gen,
+                            outgoing_event_id,
                             &event,
                             None,
+                            this_group.transactional
                         )) {
                             events.push(port_and_event);
                         };
                         // re-initialize aggr state for new window
-                        for aggr in &mut this_group.aggrs {
-                            aggr.invocable.init();
-                        }
+                        // reset transactional state for outgoing events
+                        this_group.reset();
                     }
                 }
                 _ => {
@@ -962,8 +987,14 @@ impl Operator for TrickleSelect {
                     //       | [9]    |        | [1, 2, 3, 4, 5, 6, 7, 8] // since we tilt up before we check
                     //       |        |        | emit                     // the next window we collect one too many elements
 
-                    // we need scratches for multiple window handling
+                    // we need scratches for multiple window handling, to avoid allocating on the hot path here
                     if let Some((scratch1, scratch2)) = scratches {
+                        // scratches for passing on the event id between windows to track all events contributing to an outgoing windowed event
+                        // cannot put them in AggregateScratch, as this is in tremor_script and doesnt know about EventId
+                        // cannot put AggregateScratch in TrickleSelect as it has lifetimes which we tried to avoid here
+                        let mut event_id_scratch1 = EventId::default();
+                        let mut event_id_scratch2 = EventId::default();
+
                         // gather window events
                         let mut emit_window_events = Vec::with_capacity(self.windows.len());
                         let step1_iter = self.windows.iter_mut();
@@ -1050,13 +1081,15 @@ impl Operator for TrickleSelect {
                                 // merge with scratch
                                 if !first {
                                     for (this, prev) in
-                                        this_group.aggrs.iter_mut().zip(scratch1.iter())
+                                        this_group.aggrs.iter_mut().zip(scratch1.aggregates.iter())
                                     {
                                         this.invocable.merge(&prev.invocable).map_err(|e| {
                                             let r: Option<&Registry> = None;
                                             e.into_err(prev, prev, r, &node_meta)
                                         })?;
                                     }
+                                    this_group.track_transactional(scratch1.transactional);
+                                    this_group.id.track(&event_id_scratch1);
                                 }
                                 // push event
                                 let env = Env {
@@ -1067,6 +1100,9 @@ impl Operator for TrickleSelect {
                                     recursion_limit: tremor_script::recursion_limit(),
                                 };
 
+                                let mut outgoing_event_id = self.event_id_gen.next_id();
+                                event_id_scratch1 = this_group.id.clone(); // store the id for to be tracked by the next window
+                                mem::swap(&mut outgoing_event_id, &mut this_group.id);
                                 if let Some(port_and_event) = stry!(execute_select_and_having(
                                     &stmt,
                                     &node_meta,
@@ -1074,19 +1110,21 @@ impl Operator for TrickleSelect {
                                     &local_stack,
                                     state,
                                     &env,
-                                    &mut self.event_id_gen,
+                                    outgoing_event_id,
                                     &event,
                                     None,
+                                    this_group.transactional
                                 )) {
                                     events.push(port_and_event);
                                 };
 
                                 // swap(aggrs, scratch) - store state before init to scratch as we need it for the next window
-                                std::mem::swap(scratch1, &mut this_group.aggrs);
-                                // aggrs.init()
-                                for aggr in &mut this_group.aggrs {
-                                    aggr.invocable.init();
-                                }
+                                std::mem::swap(&mut scratch1.aggregates, &mut this_group.aggrs);
+                                // store transactional state for mixing into the next window
+                                scratch1.transactional = this_group.transactional;
+
+                                // aggrs.init() and reset transactional state
+                                this_group.reset();
                             } else {
                                 // add this event to the aggr state **AFTER** emit and propagate to next windows
 
@@ -1098,6 +1136,14 @@ impl Operator for TrickleSelect {
                                     meta: &node_meta,
                                     recursion_limit: tremor_script::recursion_limit(),
                                 };
+
+                                let mut outgoing_event_id = self.event_id_gen.next_id();
+                                if !first {
+                                    // store old event_ids state from last window into scratch2
+                                    std::mem::swap(&mut event_id_scratch1, &mut event_id_scratch2);
+                                }
+                                event_id_scratch1 = this_group.id.clone(); // store the id for to be tracked by the next window
+                                mem::swap(&mut outgoing_event_id, &mut this_group.id);
                                 if let Some(port_and_event) = stry!(execute_select_and_having(
                                     &stmt,
                                     &node_meta,
@@ -1105,9 +1151,10 @@ impl Operator for TrickleSelect {
                                     &local_stack,
                                     state,
                                     &env,
-                                    &mut self.event_id_gen,
+                                    outgoing_event_id,
                                     &event,
                                     None,
+                                    this_group.transactional
                                 )) {
                                     events.push(port_and_event);
                                 };
@@ -1115,30 +1162,31 @@ impl Operator for TrickleSelect {
                                 if first {
                                     // first window
                                     //          swap(scratch, aggrs)
-                                    std::mem::swap(scratch1, &mut this_group.aggrs); // store current state before init
-                                                                                     //          aggrs.init()
-                                    for aggr in &mut this_group.aggrs {
-                                        aggr.invocable.init();
-                                    }
+                                    std::mem::swap(&mut scratch1.aggregates, &mut this_group.aggrs); // store current state before init
+                                    scratch1.transactional = this_group.transactional;
+                                    //          aggrs.init() and reset transactional state
+                                    this_group.reset();
                                 } else {
                                     // not first window
                                     // store old aggrs state from last window into scratch2
                                     std::mem::swap(scratch1, scratch2);
                                     //          swap(aggrs, scratch) // store aggrs state before init() into scratch1 for next window
-                                    std::mem::swap(&mut this_group.aggrs, scratch1);
-                                    //          aggrs.init()
-                                    for aggr in &mut this_group.aggrs {
-                                        aggr.invocable.init();
-                                    }
-                                    //          merge state from last wiundow into this one
+                                    std::mem::swap(&mut this_group.aggrs, &mut scratch1.aggregates);
+                                    scratch1.transactional = this_group.transactional;
+                                    //          aggrs.init() and reset transactional state
+                                    this_group.reset();
+
+                                    //          merge state from last window into this one
                                     for (this, prev) in
-                                        this_group.aggrs.iter_mut().zip(scratch2.iter())
+                                        this_group.aggrs.iter_mut().zip(scratch2.aggregates.iter())
                                     {
                                         this.invocable.merge(&prev.invocable).map_err(|e| {
                                             let r: Option<&Registry> = None;
                                             e.into_err(prev, prev, r, &node_meta)
                                         })?;
                                     }
+                                    this_group.track_transactional(scratch2.transactional);
+                                    this_group.id.track(&event_id_scratch2);
                                 }
                             }
 
@@ -1193,13 +1241,16 @@ impl Operator for TrickleSelect {
                                     &group_str,
                                     &group_value,
                                 ));
-                                for (this, prev) in this_group.aggrs.iter_mut().zip(scratch1.iter())
+                                for (this, prev) in
+                                    this_group.aggrs.iter_mut().zip(scratch1.aggregates.iter())
                                 {
                                     this.invocable.merge(&prev.invocable).map_err(|e| {
                                         let r: Option<&Registry> = None;
                                         e.into_err(prev, prev, r, &node_meta)
                                     })?;
                                 }
+                                this_group.track_transactional(scratch1.transactional);
+                                this_group.id.track(&event_id_scratch1);
                             }
                         }
                     } else {
@@ -1289,6 +1340,8 @@ impl Operator for TrickleSelect {
                                     meta: &node_meta,
                                     recursion_limit: tremor_script::recursion_limit(),
                                 };
+                                let mut outgoing_event_id = self.event_id_gen.next_id();
+                                mem::swap(&mut group_data.id, &mut outgoing_event_id);
                                 let executed = stry!(execute_select_and_having(
                                     &stmt,
                                     node_meta,
@@ -1296,18 +1349,17 @@ impl Operator for TrickleSelect {
                                     &local_stack,
                                     state,
                                     &env,
-                                    &mut self.event_id_gen,
+                                    outgoing_event_id,
                                     &artificial_event,
                                     None,
+                                    group_data.transactional
                                 ));
                                 if let Some(port_and_event) = executed {
                                     res.events.push(port_and_event);
                                 }
 
-                                // init
-                                for aggr in &mut group_data.aggrs {
-                                    aggr.invocable.init();
-                                }
+                                // init aggregates and reset transactional status
+                                group_data.reset();
                             }
                         }
                     }
@@ -1353,6 +1405,10 @@ impl Operator for TrickleSelect {
                         consts.group.push(group_str.clone())?;
 
                         if let Some((scratch1, scratch2)) = aggregate_scratches.as_mut() {
+                            // scratches for passing on the event id between windows to track all events contributing to an outgoing windowed event
+                            let mut event_id_scratch1 = EventId::default();
+                            let mut event_id_scratch2 = EventId::default();
+
                             // gather window events
                             let mut emit_window_events = Vec::with_capacity(self.windows.len());
                             let step1_iter = self.windows.iter_mut();
@@ -1394,14 +1450,18 @@ impl Operator for TrickleSelect {
                                     // add this event to the aggr state **BEFORE** emit and propagate to next windows
                                     // merge with scratch
                                     if !first {
-                                        for (this, prev) in
-                                            this_group.aggrs.iter_mut().zip(scratch1.iter())
+                                        for (this, prev) in this_group
+                                            .aggrs
+                                            .iter_mut()
+                                            .zip(scratch1.aggregates.iter())
                                         {
                                             this.invocable.merge(&prev.invocable).map_err(|e| {
                                                 let r: Option<&Registry> = None;
                                                 e.into_err(prev, prev, r, &node_meta)
                                             })?;
                                         }
+                                        this_group.track_transactional(scratch1.transactional);
+                                        this_group.id.track(&event_id_scratch1);
                                     }
                                     // push event
                                     let env = Env {
@@ -1412,6 +1472,9 @@ impl Operator for TrickleSelect {
                                         recursion_limit: tremor_script::recursion_limit(),
                                     };
 
+                                    let mut outgoing_event_id = self.event_id_gen.next_id();
+                                    event_id_scratch1 = this_group.id.clone(); // store event_id in a scratch for tracking it in the next window in the tilt-frame
+                                    mem::swap(&mut this_group.id, &mut outgoing_event_id);
                                     if let Some(port_and_event) = stry!(execute_select_and_having(
                                         &stmt,
                                         &node_meta,
@@ -1419,19 +1482,19 @@ impl Operator for TrickleSelect {
                                         &local_stack,
                                         state,
                                         &env,
-                                        &mut self.event_id_gen,
+                                        outgoing_event_id,
                                         &artificial_event,
                                         None,
+                                        this_group.transactional
                                     )) {
                                         res.events.push(port_and_event);
                                     };
 
                                     // swap(aggrs, scratch) - store state before init to scratch as we need it for the next window
-                                    std::mem::swap(scratch1, &mut this_group.aggrs);
-                                    // aggrs.init()
-                                    for aggr in &mut this_group.aggrs {
-                                        aggr.invocable.init();
-                                    }
+                                    std::mem::swap(&mut scratch1.aggregates, &mut this_group.aggrs);
+                                    scratch1.transactional = this_group.transactional;
+                                    // aggrs.init() and reset transactional state
+                                    this_group.reset();
                                 } else {
                                     // add this event to the aggr state **AFTER** emit and propagate to next windows
 
@@ -1443,6 +1506,17 @@ impl Operator for TrickleSelect {
                                         meta: &node_meta,
                                         recursion_limit: tremor_script::recursion_limit(),
                                     };
+
+                                    let mut outgoing_event_id = self.event_id_gen.next_id();
+                                    if !first {
+                                        // store previous window event_id in scratch2
+                                        std::mem::swap(
+                                            &mut event_id_scratch1,
+                                            &mut event_id_scratch2,
+                                        );
+                                    }
+                                    event_id_scratch1 = this_group.id.clone(); // store event id before swapping it out, for tracking in next window
+                                    mem::swap(&mut this_group.id, &mut outgoing_event_id);
                                     if let Some(port_and_event) = stry!(execute_select_and_having(
                                         &stmt,
                                         &node_meta,
@@ -1450,9 +1524,10 @@ impl Operator for TrickleSelect {
                                         &local_stack,
                                         state,
                                         &env,
-                                        &mut self.event_id_gen,
+                                        outgoing_event_id,
                                         &artificial_event,
                                         None,
+                                        this_group.transactional
                                     )) {
                                         res.events.push(port_and_event);
                                     };
@@ -1460,30 +1535,39 @@ impl Operator for TrickleSelect {
                                     if first {
                                         // first window
                                         //          swap(scratch, aggrs)
-                                        std::mem::swap(scratch1, &mut this_group.aggrs); // store current state before init
-                                                                                         //          aggrs.init()
-                                        for aggr in &mut this_group.aggrs {
-                                            aggr.invocable.init();
-                                        }
+                                        std::mem::swap(
+                                            &mut scratch1.aggregates,
+                                            &mut this_group.aggrs,
+                                        ); // store current state before init
+                                        scratch1.transactional = this_group.transactional;
+                                        //          aggrs.init() and reset transactional state
+                                        this_group.reset();
                                     } else {
                                         // not first window
                                         // store old aggrs state from last window into scratch2
                                         std::mem::swap(scratch1, scratch2);
                                         //          swap(aggrs, scratch) // store aggrs state before init() into scratch1 for next window
-                                        std::mem::swap(&mut this_group.aggrs, scratch1);
-                                        //          aggrs.init()
-                                        for aggr in &mut this_group.aggrs {
-                                            aggr.invocable.init();
-                                        }
-                                        //          merge state from last wiundow into this one
-                                        for (this, prev) in
-                                            this_group.aggrs.iter_mut().zip(scratch2.iter())
+                                        std::mem::swap(
+                                            &mut this_group.aggrs,
+                                            &mut scratch1.aggregates,
+                                        );
+                                        scratch1.transactional = this_group.transactional;
+
+                                        //          aggrs.init() and reset transactional state
+                                        this_group.reset();
+                                        //          merge state from last window into this one
+                                        for (this, prev) in this_group
+                                            .aggrs
+                                            .iter_mut()
+                                            .zip(scratch2.aggregates.iter())
                                         {
                                             this.invocable.merge(&prev.invocable).map_err(|e| {
                                                 let r: Option<&Registry> = None;
                                                 e.into_err(prev, prev, r, &node_meta)
                                             })?;
                                         }
+                                        this_group.track_transactional(scratch2.transactional);
+                                        this_group.id.track(&event_id_scratch2);
                                     }
                                 }
 
@@ -1506,13 +1590,15 @@ impl Operator for TrickleSelect {
                                         group_value,
                                     ));
                                     for (this, prev) in
-                                        this_group.aggrs.iter_mut().zip(scratch1.iter())
+                                        this_group.aggrs.iter_mut().zip(scratch1.aggregates.iter())
                                     {
                                         this.invocable.merge(&prev.invocable).map_err(|e| {
                                             let r: Option<&Registry> = None;
                                             e.into_err(prev, prev, r, &node_meta)
                                         })?;
                                     }
+                                    this_group.track_transactional(scratch1.transactional);
+                                    this_group.id.track(&event_id_scratch1);
                                 }
                             }
                         } else {
@@ -1540,11 +1626,16 @@ mod test {
 
     use super::*;
     use rental::RentalError;
-    use tremor_script::{ast::Consts, Value};
+
+    use simd_json::{json, StaticNode};
     use tremor_script::{ast::Stmt, Query};
     use tremor_script::{
         ast::{self, Ident, ImutExpr, Literal},
         path::ModulePath,
+    };
+    use tremor_script::{
+        ast::{AggregateScratch, Consts},
+        Value,
     };
 
     fn test_target<'test>() -> ast::ImutExpr<'test> {
@@ -1586,10 +1677,20 @@ mod test {
         Event {
             id: (0, 0, s).into(),
             ingest_ns: s * 1_000_000_000,
-            data: literal!({
+            data: Value::from(json!({
                "h2g2" : 42,
-            })
+            }))
             .into(),
+            ..Event::default()
+        }
+    }
+
+    fn test_event_tx(s: u64, transactional: bool, group: u64) -> Event {
+        Event {
+            id: (0, 0, s).into(),
+            ingest_ns: s + 100,
+            data: Value::from(json!({ "group": group })).into(),
+            transactional,
             ..Event::default()
         }
     }
@@ -1809,10 +1910,10 @@ mod test {
         let event = Event {
             id: (1, 1, 300).into(),
             ingest_ns: 1_000,
-            data: literal!({
+            data: Value::from(json!({
                "time" : 4,
                "g": "group"
-            })
+            }))
             .into(),
             ..Event::default()
         };
@@ -1830,10 +1931,10 @@ mod test {
         let event = Event {
             id: (1, 1, 300).into(),
             ingest_ns: 3_000,
-            data: literal!({
+            data: Value::from(json!({
                "time" : 11,
                "g": "group"
-            })
+            }))
             .into(),
             ..Event::default()
         };
@@ -1865,9 +1966,9 @@ mod test {
         let event = Event {
             id: (1, 1, 300).into(),
             ingest_ns: 2,
-            data: literal!({
+            data: Value::from(json!({
                "g": "group"
-            })
+            }))
             .into(),
             ..Event::default()
         };
@@ -1890,6 +1991,7 @@ mod test {
             r#"[{"g":"group"}]"#,
             sorted_serialize(eis.events[0].1.data.parts().0)?
         );
+        assert_eq!(false, eis.events[0].1.transactional);
         Ok(())
     }
 
@@ -1930,10 +2032,11 @@ mod test {
         let event = Event {
             id: (1, 1, 300).into(),
             ingest_ns: 300,
-            data: literal!({
+            data: Value::from(json!({
                "cat" : 42,
-            })
+            }))
             .into(),
+            transactional: true,
             ..Event::default()
         };
         eis = select.on_event(uid, "IN", &mut state, event)?;
@@ -1947,11 +2050,126 @@ mod test {
         assert_eq!(1, eis.events.len());
         let (_port, event) = eis.events.remove(0);
         assert_eq!(r#"[{"cat":42}]"#, sorted_serialize(event.data.parts().0)?);
+        assert_eq!(true, event.transactional);
 
         let mut tick5 = test_tick(499);
         eis = select.on_signal(uid, &state, &mut tick5)?;
         assert!(eis.insights.is_empty());
         assert_eq!(0, eis.events.len());
+
+        Ok(())
+    }
+
+    #[test]
+    fn test_transactional_single_window() -> Result<()> {
+        let mut op = select_stmt_from_query(
+            r#"
+            define tumbling window w2
+            with
+              size = 2
+            end;
+            select aggr::stats::count() from in[w2] into out;
+        "#,
+        )?;
+        let mut state = Value::null();
+        let event1 = test_event_tx(0, false, 0);
+        let id1 = event1.id.clone();
+        let res = op.on_event(0, "in", &mut state, event1)?;
+
+        assert!(res.events.is_empty());
+
+        let event2 = test_event_tx(1, true, 0);
+        let id2 = event2.id.clone();
+        let mut res = op.on_event(0, "in", &mut state, event2)?;
+        assert_eq!(1, res.events.len());
+        let (_, event) = res.events.pop().unwrap();
+        assert_eq!(true, event.transactional);
+        assert_eq!(true, event.id.is_tracking(&id1));
+        assert_eq!(true, event.id.is_tracking(&id2));
+
+        let event3 = test_event_tx(2, false, 0);
+        let id3 = event3.id.clone();
+        let res = op.on_event(0, "in", &mut state, event3)?;
+        assert!(res.events.is_empty());
+
+        let event4 = test_event_tx(3, false, 0);
+        let id4 = event4.id.clone();
+        let mut res = op.on_event(0, "in", &mut state, event4)?;
+        assert_eq!(1, res.events.len());
+        let (_, event) = res.events.pop().unwrap();
+        assert_eq!(false, event.transactional);
+        assert_eq!(true, event.id.is_tracking(&id3));
+        assert_eq!(true, event.id.is_tracking(&id4));
+
+        Ok(())
+    }
+
+    #[test]
+    fn test_transactional_multiple_windows() -> Result<()> {
+        let mut op = select_stmt_from_query(
+            r#"
+            define tumbling window w2_1
+            with
+              size = 2
+            end;
+            define tumbling window w2_2
+            with
+              size = 2
+            end;
+            select aggr::win::collect_flattened(event) from in[w2_1, w2_2] group by set(event["group"]) into out;
+        "#,
+        )?;
+        let mut state = Value::null();
+        let event1 = test_event_tx(0, true, 0);
+        let id1 = event1.id.clone();
+        let res = op.on_event(0, "in", &mut state, event1)?;
+        assert_eq!(0, res.len());
+
+        let event2 = test_event_tx(1, false, 1);
+        let id2 = event2.id.clone();
+        let res = op.on_event(0, "in", &mut state, event2)?;
+        assert_eq!(0, res.len());
+
+        let event3 = test_event_tx(2, false, 0);
+        let id3 = event3.id.clone();
+        let mut res = op.on_event(0, "in", &mut state, event3)?;
+        assert_eq!(1, res.len());
+        let (_, event) = res.events.pop().unwrap();
+        assert_eq!(true, event.transactional);
+        assert_eq!(true, event.id.is_tracking(&id1));
+        assert_eq!(true, event.id.is_tracking(&id3));
+
+        let event4 = test_event_tx(3, false, 1);
+        let id4 = event4.id.clone();
+        let mut res = op.on_event(0, "in", &mut state, event4)?;
+        assert_eq!(1, res.len());
+        let (_, event) = res.events.remove(0);
+        assert_eq!(false, event.transactional);
+        assert_eq!(true, event.id.is_tracking(&id2));
+        assert_eq!(true, event.id.is_tracking(&id4));
+
+        let event5 = test_event_tx(4, false, 0);
+        let id5 = event5.id.clone();
+        let res = op.on_event(0, "in", &mut state, event5)?;
+        assert_eq!(0, res.len());
+
+        let event6 = test_event_tx(5, false, 0);
+        let id6 = event6.id.clone();
+        let mut res = op.on_event(0, "in", &mut state, event6)?;
+        assert_eq!(2, res.len());
+
+        // first event from event5 and event6 - none of the source events are transactional
+        let (_, event_w1) = res.events.remove(0);
+        assert_eq!(false, event_w1.transactional);
+        assert_eq!(true, event_w1.id.is_tracking(&id5));
+        assert_eq!(true, event_w1.id.is_tracking(&id6));
+
+        let (_, event_w2) = res.events.remove(0);
+        assert_eq!(true, event_w2.transactional);
+        assert_eq!(true, event_w2.id.is_tracking(&id1));
+        assert_eq!(true, event_w2.id.is_tracking(&id3));
+        assert_eq!(true, event_w2.id.is_tracking(&id5));
+        assert_eq!(true, event_w2.id.is_tracking(&id6));
 
         Ok(())
     }
@@ -2253,7 +2471,10 @@ mod test {
 
     fn test_select_stmt(stmt: tremor_script::ast::Select) -> tremor_script::ast::Stmt {
         let aggregates = vec![];
-        let aggregate_scratches = Some((aggregates.clone(), aggregates.clone()));
+        let aggregate_scratches = Some((
+            AggregateScratch::new(aggregates.clone()),
+            AggregateScratch::new(aggregates.clone()),
+        ));
         ast::Stmt::Select(SelectStmt {
             stmt: Box::new(stmt),
             aggregates,
@@ -2387,7 +2608,7 @@ mod test {
         Ok(())
     }
 
-    fn json_event(ingest_ns: u64, payload: Value<'static>) -> Event {
+    fn json_event(ingest_ns: u64, payload: OwnedValue) -> Event {
         Event {
             id: (0, 0, ingest_ns).into(),
             ingest_ns,
@@ -2423,7 +2644,7 @@ mod test {
             other => return Err(format!("Didnt get a window decl, got: {:?}", other).into()),
         };
         let mut params = halfbrown::HashMap::with_capacity(1);
-        params.insert("size".to_string(), Value::from(3_u64));
+        params.insert("size".to_string(), Value::Static(StaticNode::U64(3)));
         let interval = window_decl
             .params
             .get("interval")
@@ -2437,7 +2658,7 @@ mod test {
             Some(&window_decl),
             &stmt,
         );
-        let json1 = literal!({
+        let json1 = json!({
             "timestamp": 1_000_000_000
         });
         assert_eq!(
@@ -2448,7 +2669,7 @@ mod test {
             },
             window.on_event(&json_event(1, json1))?
         );
-        let json2 = literal!({
+        let json2 = json!({
             "timestamp": 1_999_999_999
         });
         assert_eq!(
@@ -2459,7 +2680,7 @@ mod test {
             },
             window.on_event(&json_event(2, json2))?
         );
-        let json3 = literal!({
+        let json3 = json!({
             "timestamp": 2_000_000_000
         });
         // ignoring on_tick as we have a script
@@ -2523,7 +2744,7 @@ mod test {
                 include: false,
                 emit: false
             },
-            window.on_event(&json_event(101, Value::object()))?
+            window.on_event(&json_event(101, json!({})))?
         );
         assert_eq!(
             WindowEvent {
@@ -2585,7 +2806,7 @@ mod test {
                 include: false,
                 emit: false
             },
-            window.on_event(&json_event(101, Value::object()))?
+            window.on_event(&json_event(101, json!({})))?
         );
         assert_eq!(
             WindowEvent {